% Generated by build script
\name{bls_catalog}
\alias{bls_catalog}
\title{Fetch series catalog metadata}
\usage{
bls_catalog()
}
\arguments{
\item{series}{Character vector of series IDs}
\item{key}{BLS API registration key (required for catalog)}
}
\value{tibble: series_id, series_title, survey_name, survey_abbreviation,}
\description{Fetch series catalog metadata}
