% Generated by build script
\name{bls_earnings}
\alias{bls_earnings}
\title{Average Hourly Earnings (private sector, seasonally adjusted)}
\usage{
bls_earnings()
}
\arguments{
\item{start_year}{Start year}
\item{end_year}{End year}
\item{key}{Optional API key}
}
\value{tibble: series_id, year, period, period_name, value, date}
\description{Average Hourly Earnings (private sector, seasonally adjusted)}
