% Generated by build script
\name{bls_eci}
\alias{bls_eci}
\title{Employment Cost Index (total compensation, private)}
\usage{
bls_eci()
}
\arguments{
\item{start_year}{Start year}
\item{end_year}{End year}
\item{key}{Optional API key}
}
\value{tibble: series_id, year, period, period_name, value, date}
\description{Employment Cost Index (total compensation, private)}
