% Generated by build script
\name{bls_jolts}
\alias{bls_jolts}
\title{Job Openings (JOLTS, total nonfarm)}
\usage{
bls_jolts()
}
\arguments{
\item{start_year}{Start year}
\item{end_year}{End year}
\item{key}{Optional API key}
}
\value{tibble: series_id, year, period, period_name, value, date}
\description{Job Openings (JOLTS, total nonfarm)}
