% Generated by build script
\name{bls_payrolls}
\alias{bls_payrolls}
\title{Total Nonfarm Payrolls (seasonally adjusted)}
\usage{
bls_payrolls()
}
\arguments{
\item{start_year}{Start year}
\item{end_year}{End year}
\item{key}{Optional API key}
}
\value{tibble: series_id, year, period, period_name, value, date}
\description{Total Nonfarm Payrolls (seasonally adjusted)}
