% Generated by build script
\name{bls_ppi}
\alias{bls_ppi}
\title{Producer Price Index (all commodities)}
\usage{
bls_ppi()
}
\arguments{
\item{start_year}{Start year}
\item{end_year}{End year}
\item{key}{Optional API key}
}
\value{tibble: series_id, year, period, period_name, value, date}
\description{Producer Price Index (all commodities)}
