% Generated by build script
\name{bls_unemployment}
\alias{bls_unemployment}
\title{Unemployment Rate (seasonally adjusted)}
\usage{
bls_unemployment()
}
\arguments{
\item{start_year}{Start year}
\item{end_year}{End year}
\item{key}{Optional API key}
}
\value{tibble: series_id, year, period, period_name, value, date}
\description{Unemployment Rate (seasonally adjusted)}
